#1 restructure using classes
